BZ #715103 aeolus-image push w/ build id not implemented / fails
https://bugzilla.redhat.com/show_bug.cgi?id=715103

This started as a patch to tidy up tests and exits, but turned out
to also take care of this BZ.  The user should now get appropriate
feedback when calling this not-yet-implemented combination.

This should result in no other functional changes, just removes some
duplication in the tests, and give a non-zero exit code for
error cases, which we should not have had exit with 0 previously.