BZ# 768509: aeolus-cli rename to aeolus
https://bugzilla.redhat.com/show_bug.cgi?id=768509

If we rename aeolus-cli > aeolus, then it doesn't make much
sense to have the subcommands be of the form 'aeolus-cli-{command}'.
This patch sets up the first convenience executable to allow power
users to invoke the image manement tooling directly.  This pattern
can then be repeated as other parts of aeolus get their own cli
interface (aeolus-deployable, for example).