BZ862563 - "aeolus-image list --providerimages" fails
https://bugzilla.redhat.com/show_bug.cgi?id=862563

The provider account API has changed to return 'shallow' objects when
querying the collection (e.g. only id and href to the 'full'
resource).  We need to chase all the hrefs in order to get the labels
and provider types.

Also, in the provider account API, the 'name' attribute was changed to
'label', so adjust accordingly.